requirements.txt
# local package
-e .

# external requirements
click~=7.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.0.3
setuptools~=46.4.0
numpy~=1.18.4
scikit-learn~=0.23.1
src~=0.1.0
xgboost~=1.0.2